feat: creation of local git management
first approach

+++ src/endpoints/github.py
@@ -1,11 +1,13 @@
 from fastapi import Depends, HTTPException, Security, UploadFile, Form, File, exceptions, Request, responses, status
 from typing import Union, Annotated, Optional
 from pydantic import BaseModel
+from src.my_types import Autor
 
 import src.configs as configs
 from src.logs import logging
 from src.service import MainService
 from src.endpoints.oauth import decode_token, OAuthCallbackResponse
+import src.services.repomanager as repomanager
 
 from fastapi import APIRouter
 import os
@@ -13,6 +15,7 @@ import os
 router = APIRouter()
 # service = MainService(configs.config.github_token, configs.config.repo_owner, configs.config)
 service = None
+REPOS_PATH = "repos"
 
 class BasicResponse(BaseModel):
     message: str
@@ -30,28 +33,48 @@ async def get_data(payload: OAuthCallbackResponse = Security(decode_token)):
     return {"message": "Here is your data", "user": payload}
 
 @router.post("/uploadfiles/")
-async def upload_files(repository: str = "", path: str = "", files: list[UploadFile] = File(...), payload: OAuthCallbackResponse = Security(decode_token)):
-    """
-    Upload files to the specified repository and path, and create PR on this branch.
+async def upload_files(repository: str = Form(...), path: str = Form(...), files: list[UploadFile] = File(...), payload: OAuthCallbackResponse = Security(decode_token)):
+    """Upload files to the specified repository and path, and create PR on this branch.
 
-    :param repository: Local GitHub repository path.
-    :param path: Path inside the root of the repository.
-    :param files: List of files to upload.
-    :return: Response message.
+    Args
+    ----
+    repository: str
+        Local GitHub repository path.
+    path: str
+        Path inside the root of the repository.
+
+    Returns
+    -------
+    OAuthCallbackResponse
     """
     # Check if repository and path are provided
     if not repository or not path:
         raise HTTPException(status_code=400, detail="Repository and path must be provided.")
 
-    upload_dir = os.path.join(repository, path)
+    repository_path = os.path.join(REPOS_PATH, repository)
+    if not repomanager.check_repo_exists(repository_path):
+        raise HTTPException(status_code=400, detail=f"Requested repository {repository} does not exist.")
+
+    print(payload, "------------")
+    author: Autor = Autor(name=payload.username, email=payload.email)
+    branch_name = repomanager.create_branch_name(author)
+    repomanager.move_or_create_to_branch(repository_path, branch_name)
+    upload_dir = os.path.join(repository_path, path)
     # Define the directory to save files
-    upload_dir = "uploads"
     os.makedirs(upload_dir, exist_ok=True)
 
-    for file in files:
+    files_to_commit = [""] * len(files)
+    for i, file in enumerate(files):
         file_path = os.path.join(upload_dir, file.filename)
         with open(file_path, "wb") as buffer:
             buffer.write(await file.read())
+        files_to_commit[i] = file_path
+
+    commit_files = repomanager.commit_files(repository_path, files_to_commit, author)
+    repomanager.move_or_create_to_branch(repository_path, "main")
+
+    if not commit_files:
+        raise HTTPException(status_code=500, detail="Failed to commit files to the repository.")
 
     return {"detail": f"{len(files)} files uploaded successfully"}
